refactor: making a reusable create server function
