+++ src/routes/HomeRoute.test.js
@@ -1,35 +1,23 @@
 import { render, screen } from "@testing-library/react";
-import { setupServer } from "msw/node";
-import { rest } from "msw";
 
 import HomeRoute from "./HomeRoute";
 import { MemoryRouter } from "react-router";
+import { createServer } from "../test/server";
 
-const handlers = [
-  rest.get("/api/repositories", (req, res, ctx) => {
-    const language = req.url.searchParams.get("q").split("language:")[1];
-
-    return res(
-      ctx.json({
+createServer([
+  {
+    path: "/api/repositories",
+    res: (req, res, ctx) => {
+      const language = req.url.searchParams.get("q").split("language:")[1];
+      return {
         items: [
           { id: 1, full_name: `${language}_one` },
           { id: 2, full_name: `${language}_two` },
         ],
-      })
-    );
-  }),
-];
-const server = setupServer(...handlers);
-
-beforeAll(() => {
-  server.listen();
-});
-afterEach(() => {
-  server.resetHandlers();
-});
-afterAll(() => {
-  server.close();
-});
+      };
+    },
+  },
+]);
 
 test("Should renders two links for each language", async () => {
   render(
